Rename CardSerialization to cardSerialization
It's only exporting functions, not a class
