+++ tests/cards/underworld/PriceWars.spec.ts
@@ -2,7 +2,7 @@ import {expect} from 'chai';
 import {PriceWars} from '../../../src/server/cards/underworld/PriceWars';
 import {testGame} from '../../TestGame';
 import {cast, forceGenerationEnd, runAllActions} from '../../TestingUtils';
-import {deserializeProjectCard, serializeProjectCard} from '../../../src/server/cards/CardSerialization';
+import {deserializeProjectCard, serializeProjectCard} from '../../../src/server/cards/cardSerialization';
 
 describe('PriceWars', () => {
   it('canPlay', () => {
